android/app/src/main/java/com/crab/common/mapper/ItemDtoToItemEntityMapper.java
package com.crab.common.mapper;

import com.crab.models.dto.ItemDto;
import com.crab.models.entities.ItemEntity;
import com.crab.utils.Constants;

import org.bson.types.ObjectId;
import org.mapstruct.Mapper;
import org.mapstruct.Mapping;
import org.mapstruct.Named;

import java.text.ParseException;
import java.text.SimpleDateFormat;
import java.util.Date;
import java.util.Calendar;

@Mapper
public abstract class ItemDtoToItemEntityMapper{
    public abstract ItemEntity itemDtoToItemEntity(ItemDto itemDto);
}
